HSEARCH-3118 Extract the thread-safe shutdown feature of Elasticsearch work orchestrators to an abstract class in the engine
So we can reuse it in the Lucene backend.
