+++ backend/elasticsearch/src/main/java/org/hibernate/search/backend/elasticsearch/orchestration/impl/AbstractElasticsearchWorkOrchestrator.java
@@ -6,41 +6,24 @@
  */
 package org.hibernate.search.backend.elasticsearch.orchestration.impl;
 
-import java.lang.invoke.MethodHandles;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.concurrent.CompletableFuture;
-import java.util.concurrent.locks.ReadWriteLock;
-import java.util.concurrent.locks.ReentrantReadWriteLock;
 
-import org.hibernate.search.backend.elasticsearch.logging.impl.Log;
 import org.hibernate.search.backend.elasticsearch.work.impl.ElasticsearchWork;
+import org.hibernate.search.engine.backend.orchestration.spi.AbstractWorkOrchestrator;
 import org.hibernate.search.engine.backend.orchestration.spi.BatchingExecutor;
 import org.hibernate.search.util.common.impl.Futures;
-import org.hibernate.search.util.common.logging.impl.LoggerFactory;
 
 /**
- * An abstract base for {@link ElasticsearchWorkOrchestratorImplementor} implementations,
- * implementing a thread-safe shutdown.
- *
- * @author Yoann Rodiere
+ * An abstract base for {@link ElasticsearchWorkOrchestratorImplementor} implementations.
  */
 abstract class AbstractElasticsearchWorkOrchestrator
-		implements ElasticsearchWorkOrchestratorImplementor, AutoCloseable {
+		extends AbstractWorkOrchestrator<AbstractElasticsearchWorkOrchestrator.ElasticsearchWorkSet>
+		implements ElasticsearchWorkOrchestratorImplementor {
 
-	private static final Log log = LoggerFactory.make( Log.class, MethodHandles.lookup() );
-
-	private final String name;
-
-	private boolean open = true; // Guarded by shutdownLock
-	private final ReadWriteLock shutdownLock = new ReentrantReadWriteLock();
-
-	protected AbstractElasticsearchWorkOrchestrator(String name) {
-		this.name = name;
-	}
-
-	protected final String getName() {
-		return name;
+	AbstractElasticsearchWorkOrchestrator(String name) {
+		super( name );
 	}
 
 	@Override
@@ -55,46 +38,6 @@ abstract class AbstractElasticsearchWorkOrchestrator
 		CompletableFuture<T> future = new CompletableFuture<>();
 		submit( new ElasticsearchSingleWorkSet<>( work, future ) );
 		return future;
-	}
-
-	@Override
-	public void close() {
-		shutdownLock.writeLock().lock();
-		try {
-			if ( !open ) {
-				return;
-			}
-			open = false;
-			doClose();
-		}
-		finally {
-			shutdownLock.writeLock().unlock();
-		}
-	}
-
-	protected abstract void doSubmit(ElasticsearchWorkSet workSet) throws InterruptedException;
-
-	protected abstract void doClose();
-
-	void submit(ElasticsearchWorkSet workSet) {
-		if ( !shutdownLock.readLock().tryLock() ) {
-			// The orchestrator is shutting down: abort.
-			throw log.orchestratorShutDownBeforeSubmittingWorkset( name );
-		}
-		try {
-			if ( !open ) {
-				// The orchestrator has shut down: abort.
-				throw log.orchestratorShutDownBeforeSubmittingWorkset( name );
-			}
-			doSubmit( workSet );
-		}
-		catch (InterruptedException e) {
-			Thread.currentThread().interrupt();
-			throw log.threadInterruptedWhileSubmittingWorkset( name );
-		}
-		finally {
-			shutdownLock.readLock().unlock();
-		}
 	}
 
 	interface ElasticsearchWorkSet extends BatchingExecutor.Task<ElasticsearchWorkOrchestrationStrategy> {
